+++ src/main/java/uz/pdp/clickup/service/TasksUsersServiceImpl.java
@@ -3,10 +3,12 @@ package uz.pdp.clickup.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import uz.pdp.clickup.entity.Task;
+import uz.pdp.clickup.entity.TaskHistory;
 import uz.pdp.clickup.entity.TasksUsers;
 import uz.pdp.clickup.entity.User;
 import uz.pdp.clickup.payload.ApiResponse;
 import uz.pdp.clickup.payload.TasksUsersDto;
+import uz.pdp.clickup.repository.TaskHistoryRepository;
 import uz.pdp.clickup.repository.TaskRepository;
 import uz.pdp.clickup.repository.TasksUsersRepository;
 import uz.pdp.clickup.repository.UserRepository;
@@ -27,6 +29,9 @@ public class TasksUsersServiceImpl implements TasksUsersService {
 
     @Autowired
     TasksUsersRepository tasksUsersRepository;
+
+    @Autowired
+    TaskHistoryRepository taskHistoryRepository;
 
     @Override
     public ApiResponse addTasksUsers(TasksUsersDto tasksUsersDto) {
@@ -49,10 +54,25 @@ public class TasksUsersServiceImpl implements TasksUsersService {
                         optionalTask.get(),
                         optionalUser.get()
                 ));
+
+
+
+                /*
+                 * Task history
+                 * */
+                taskHistoryRepository.save(new TaskHistory(
+                        optionalTask.get(),
+                        "TasUsers",
+                        "",
+                        "add",
+                        "Users added to task"
+                ));
             }
         }
 
         tasksUsersRepository.saveAll(tasksUsersList);
+
+
 
         return new ApiResponse("TasksUsers saved!", true);
     }
@@ -60,7 +80,23 @@ public class TasksUsersServiceImpl implements TasksUsersService {
     @Override
     public ApiResponse deleteTasksUsers(Long id) {
         try {
+
+            Optional<TasksUsers> optionalTasksUsers = tasksUsersRepository.findById(id);
+            if (optionalTasksUsers.isEmpty())
+                return new ApiResponse("TaskUsers not found!", false);
+
             tasksUsersRepository.deleteById(id);
+
+            /*
+             * Task history
+             * */
+            taskHistoryRepository.save(new TaskHistory(
+                    optionalTasksUsers.get().getTask(),
+                    "TasUsers",
+                    "add",
+                    "remove",
+                    "User deleted from task"
+            ));
             return new ApiResponse("TasksUsers deleted!", true);
         }catch (Exception e) {
             return new ApiResponse("Error", false);
